+++ src/tool/poetry.rs
@@ -48,7 +48,7 @@ fn is_poetry_pyproject(content: &str) -> bool {
 
 /// `poetry install`
 pub(crate) fn install_cmd() -> Command {
-    let mut c = Command::new("poetry");
+    let mut c = super::program::command("poetry");
     c.arg("install");
     c
 }